+++ CHANGELOG.md
@@ -4,6 +4,8 @@ TODO add summary
 
 ## MINOR CHANGES
 
+* `Nextflow VDSL3`: Add profiles to the Nextflow Config file when the main script is a `NextflowScript` (#408).
+
 * `Scripts`: Remove the `entrypoint` parameter for all script types except `NextflowScript`. All these scripts had to check individually whether the parameter was unset, now it can be done in the `Script` apply method. (#409)
 
 # Viash 0.7.3 (2023-04-19): Minor bug fixes in documentation and config view